feat: alternative styles for mobile devices
