+++ components/common/FormPage.tsx
@@ -1,4 +1,4 @@
-import { Box } from "@chakra-ui/react"
+import { Box, useMediaQuery } from "@chakra-ui/react"
 import React from "react"
 import Header from "./Header"
 
@@ -8,10 +8,12 @@ interface IFormPage {
 }
 
 const FormPage = (props: IFormPage): JSX.Element => {
+    const [isLarge] = useMediaQuery("(min-width: 900px)")
     const { header, children } = props
+
     return (
-        <Box>
-            <Header type="main">{header}</Header>
+        <Box minWidth={isLarge ? "50vw" : "100vw"}>
+            {isLarge && <Header type="main">{header}</Header>}
             <Box backgroundColor="whitesmoke" minH="100vh">
                 {children}
             </Box>
