+++ components/common/Layout.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from "react"
-import Header from "./Header"
+import { TopHeader } from "./Header"
 import Main from "./Main"
 import { Flex } from "@chakra-ui/react"
 
@@ -9,13 +9,14 @@ interface LayoutProps {
 
 const Layout = (props: LayoutProps): JSX.Element => {
     const { children } = props
+
     return (
         <Flex
             flexDirection="column"
             minHeight="100vh"
             background="linear-gradient(#9f9f9f, #efefef)"
         >
-            <Header type="top" />
+            <TopHeader />
             <Main>{children}</Main>
         </Flex>
     )
