+++ components/table/ReportTable/DateInput.tsx
@@ -11,8 +11,8 @@ import {
 import { User } from "firebase/auth"
 
 interface IDateInput {
-    startDate: string
-    endDate: string
+    startDate: string | null
+    endDate: string | null
     setStartDate: React.Dispatch<string>
     setEndDate: React.Dispatch<string>
     user: User
@@ -24,14 +24,14 @@ const DateInput = ({
     setStartDate,
     setEndDate,
 }: IDateInput): JSX.Element => {
-    const isInvalid = endDate < startDate || !endDate || !startDate
+    const isInvalid = !endDate || !startDate || endDate < startDate
 
     const handleStartDateChange = (
         event: React.FormEvent<HTMLInputElement>
     ) => {
         if (
-            endDate >= event.currentTarget.value &&
             endDate &&
+            endDate >= event.currentTarget.value &&
             event.currentTarget.value
         ) {
             setStartDate(event.currentTarget.value)
@@ -40,9 +40,9 @@ const DateInput = ({
 
     const handleEndDateChange = (event: React.FormEvent<HTMLInputElement>) => {
         if (
+            startDate &&
             event.currentTarget.value >= startDate &&
-            event.currentTarget.value &&
-            startDate
+            event.currentTarget.value
         ) {
             setEndDate(event.currentTarget.value)
         }
@@ -57,9 +57,9 @@ const DateInput = ({
                 <FormControl isInvalid={isInvalid && !isUntouched}>
                     <Input
                         type={"date"}
-                        value={startDate}
+                        value={startDate ?? undefined}
                         onChange={handleStartDateChange}
-                    ></Input>
+                    />
                     {isInvalid ? (
                         <FormErrorMessage>Invalid dates</FormErrorMessage>
                     ) : (
@@ -69,9 +69,9 @@ const DateInput = ({
                 <FormControl isInvalid={isInvalid && !isUntouched}>
                     <Input
                         type={"date"}
-                        value={endDate}
+                        value={endDate ?? undefined}
                         onChange={handleEndDateChange}
-                    ></Input>
+                    />
                     {isInvalid ? (
                         <FormErrorMessage>Invalid dates</FormErrorMessage>
                     ) : (
