+++ package.json
@@ -31,8 +31,8 @@
         "@emotion/react": "^11.10.0",
         "@emotion/styled": "^11.10.0",
         "@swc/core": "^1.2.241",
-        "csv-parse": "^5.3.2",
         "@types/pdfmake": "^0.2.2",
+        "csv-parse": "^5.3.2",
         "firebase": "^9.9.3",
         "firebase-admin": "^11.0.1",
         "framer-motion": "^7.2.0",
