+++ pages/index.tsx
@@ -1,7 +1,6 @@
 import React, { useContext } from "react"
 import type { NextPage } from "next"
 import { UserContext } from "@/lib/contexts/FirebaseAuthContext"
-import { TimesheetEntryEditor } from "@/components/editor/TimesheetEntryEditor"
 import ErrorAlert from "@/components/common/ErrorAlert"
 import Loading from "@/components/common/Loading"
 import {
@@ -11,6 +10,7 @@ import {
 } from "@/lib/hooks/useList"
 import FormPage from "@/components/common/FormPage"
 import { Box } from "@chakra-ui/react"
+import TimesheetEntryEditor from "@/components/editor/TimesheetEntryEditor"
 
 interface IndexPageProps {
     email: string
