+++ styles/Calendar.css
@@ -1,12 +1,11 @@
 .react-calendar {
-    width: 40rem;
     max-width: 100%;
     background: white;
     font-family: Arial, Helvetica, sans-serif;
     line-height: 1.125em;
 }
 .react-calendar--doubleView {
-    width: 700px;
+    max-width: 100%;
 }
 .react-calendar--doubleView .react-calendar__viewContainer {
     display: flex;
